Add files via upload
The first file is a magazine that describes the Skyline High School Mountain Bike Team. I used two photos that were take an the first race of the 2020 season. The second image is supposed to represent a couch potato and I also added in a Pringles can for some irony. The last picture is a picture of a computer in a computer showing two dogs working.